+++ nexus/mgs-updates/src/test_util/sp_test_state.rs
@@ -56,7 +56,7 @@ impl SpTestState {
     pub async fn load(
         mgs_client: &gateway_client::Client,
         sp_type: SpType,
-        sp_slot: u32,
+        sp_slot: u16,
     ) -> Result<SpTestState, GatewayClientError> {
         let caboose_sp_active = mgs_client
             .sp_component_caboose_get(
